Change the folder structure with a new Testing parent folder.
