+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Testing/MsTestV2/UnitTestResultOutcome.cs
@@ -1,4 +1,4 @@
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.MsTestV2
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.MsTestV2
 {
     /// <summary>
     /// UnitTestResult Outcome
