+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Testing/MsTestV2/UnitTestResultStruct.cs
@@ -1,7 +1,7 @@
 using System;
 using Datadog.Trace.DuckTyping;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.MsTestV2
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.MsTestV2
 {
     /// <summary>
     /// UnitTestResult ducktype struct
