+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Testing/NUnit/IMethodInfo.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.NUnit
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.NUnit
 {
     /// <summary>
     /// DuckTyping interface for NUnit.Framework.Interfaces.IMethodInfo
