+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Testing/NUnit/IPropertyBag.cs
@@ -1,7 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.NUnit
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.NUnit
 {
     /// <summary>
     /// DuckTyping interface for NUnit.Framework.Interfaces.IPropertyBag
