+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Testing/NUnit/ITest.cs
@@ -1,4 +1,4 @@
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.NUnit
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.NUnit
 {
     /// <summary>
     /// DuckTyping interface for NUnit.Framework.Internal.Test
