+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Testing/XUnit/TestInvokerStruct.cs
@@ -2,13 +2,13 @@ using System;
 using System.Reflection;
 using Datadog.Trace.DuckTyping;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.XUnit
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.XUnit
 {
     /// <summary>
-    /// TestRunner`1 structure
+    /// TestInvoker`1 structure
     /// </summary>
     [DuckCopy]
-    public struct TestRunnerStruct
+    public struct TestInvokerStruct
     {
         /// <summary>
         /// Test class Type
@@ -34,10 +34,5 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.XUnit
         /// Exception aggregator
         /// </summary>
         public IExceptionAggregator Aggregator;
-
-        /// <summary>
-        /// Skip reason
-        /// </summary>
-        public string SkipReason;
     }
 }
